+++ packages/camera/camera_platform_interface/lib/src/platform_interface/camera_platform.dart
@@ -113,13 +113,11 @@ abstract class CameraPlatform extends PlatformInterface {
   /// Implementations for this:
   /// - Should support all 4 orientations.
   Stream<DeviceOrientationChangedEvent> onDeviceOrientationChanged() {
-    throw UnimplementedError(
-        'onDeviceOrientationChanged() is not implemented.');
+    throw UnimplementedError('onDeviceOrientationChanged() is not implemented.');
   }
 
   /// Locks the capture orientation.
-  Future<void> lockCaptureOrientation(
-      int cameraId, DeviceOrientation orientation) {
+  Future<void> lockCaptureOrientation(int cameraId, DeviceOrientation orientation) {
     throw UnimplementedError('lockCaptureOrientation() is not implemented.');
   }
 
@@ -156,8 +154,7 @@ abstract class CameraPlatform extends PlatformInterface {
   /// Please see [VideoCaptureOptions] for documentation on the
   /// configuration options.
   Future<void> startVideoCapturing(VideoCaptureOptions options) {
-    return startVideoRecording(options.cameraId,
-        maxVideoDuration: options.maxDuration);
+    return startVideoRecording(options.cameraId, maxVideoDuration: options.maxDuration);
   }
 
   /// Stops the video recording and returns the file where it was saved.
@@ -185,8 +182,7 @@ abstract class CameraPlatform extends PlatformInterface {
   ///
   // TODO(bmparr): Add options to control streaming settings (e.g.,
   // resolution and FPS).
-  Stream<CameraImageData> onStreamedFrameAvailable(int cameraId,
-      {CameraImageStreamOptions? options}) {
+  Stream<CameraImageData> onStreamedFrameAvailable(int cameraId, {CameraImageStreamOptions? options}) {
     throw UnimplementedError('onStreamedFrameAvailable() is not implemented.');
   }
 
@@ -285,8 +281,7 @@ abstract class CameraPlatform extends PlatformInterface {
 
   /// Sets the active camera while recording.
   Future<void> setDescriptionWhileRecording(CameraDescription description) {
-    throw UnimplementedError(
-        'setDescriptionWhileRecording() is not implemented.');
+    throw UnimplementedError('setDescriptionWhileRecording() is not implemented.');
   }
 
   /// Returns a widget showing a live camera preview.
@@ -301,6 +296,6 @@ abstract class CameraPlatform extends PlatformInterface {
 
   /// Sets the output format for the selected camera.
   Future<void> setImageFileFormat(int cameraId, ImageFileFormat format) {
-    throw UnimplementedError('setFileFormat() is not implemented.');
+    throw UnimplementedError('setImageFileFormat() is not implemented.');
   }
 }