+++ packages/camera/camera/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## NEXT
 
-* Updates minimum supported SDK version to Flutter 3.10/Dart 3.
+* Updates minimum supported SDK version to Flutter 3.10/Dart 3.0.
 
 ## 0.10.5+5