+++ packages/camera/camera_avfoundation/CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 0.9.14
 
 * Adds support to HEIF format.
-* Updates minimum supported SDK version to Flutter 3.10/Dart 3.
+* Updates minimum supported SDK version to Flutter 3.10/Dart 3.0.
 
 ## 0.9.13+7