+++ packages/camera/camera_platform_interface/CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 2.7.0
 
 * Adds support to HEIF format for iOS.
-* Updates minimum supported SDK version to Flutter 3.10/Dart 3.
+* Updates minimum supported SDK version to Flutter 3.10/Dart 3.0.
 
 ## 2.6.0